Renamed the old WorldManager class to"OldWorldManager" and created a new "WorldManager" class
