+++ NitroxLauncher/Pages/OptionPage.xaml.cs
@@ -19,7 +19,7 @@ namespace NitroxLauncher.Pages
         public OptionPage()
         {
             InitializeComponent();
-            SaveFileLocationTextblock.Text = WorldManager.SavesFolderDir;
+            SaveFileLocationTextblock.Text = OldWorldManager.SavesFolderDir;
 
             ArgumentsTextbox.Text = SubnauticaLaunchArguments;
             if (SubnauticaLaunchArguments != LauncherConfig.DEFAULT_LAUNCH_ARGUMENTS)
@@ -104,7 +104,7 @@ namespace NitroxLauncher.Pages
 
         private void OnViewFolder_Click(object sender, RoutedEventArgs e)
         {
-            Process.Start(WorldManager.SavesFolderDir)?.Dispose();
+            Process.Start(OldWorldManager.SavesFolderDir)?.Dispose();
         }
     }
 }
