+++ NitroxServer/Server.cs
@@ -62,7 +62,7 @@ namespace NitroxServer
             StringBuilder builder = new("\n");
             if (viewerPerms is Perms.CONSOLE)
             {
-                builder.AppendLine($" - Save location: {Path.Combine(WorldManager.SavesFolderDir, serverConfig.SaveName)}");
+                builder.AppendLine($" - Save location: {Path.Combine(OldWorldManager.SavesFolderDir, serverConfig.SaveName)}");
             }
             builder.AppendLine($"""
              - Aurora's state: {world.StoryManager.GetAuroraStateSummary()}
@@ -84,7 +84,7 @@ namespace NitroxServer
             string saveDir = null;
             foreach (string arg in Environment.GetCommandLineArgs())
             {
-                if (arg.StartsWith(WorldManager.SavesFolderDir, StringComparison.OrdinalIgnoreCase) && Directory.Exists(arg))
+                if (arg.StartsWith(OldWorldManager.SavesFolderDir, StringComparison.OrdinalIgnoreCase) && Directory.Exists(arg))
                 {
                     saveDir = arg;
                     break;
@@ -93,7 +93,7 @@ namespace NitroxServer
             if (saveDir == null)
             {
                 // Check if there are any save files
-                WorldManager.Listing[] worldList = WorldManager.GetSaves().ToArray();
+                OldWorldManager.Listing[] worldList = OldWorldManager.GetSaves().ToArray();
                 if (worldList.Any())
                 {
                     // Get last save file used
@@ -113,7 +113,7 @@ namespace NitroxServer
                 else
                 {
                     // Create new save file
-                    saveDir = Path.Combine(WorldManager.SavesFolderDir, "My World");
+                    saveDir = Path.Combine(OldWorldManager.SavesFolderDir, "My World");
                     Directory.CreateDirectory(saveDir);
                     ServerConfig serverConfig = ServerConfig.Load(saveDir);
                     Log.Debug($"No save file was found, creating a new one...");
@@ -133,7 +133,7 @@ namespace NitroxServer
 
             IsSaving = true;
 
-            bool savedSuccessfully = worldPersistence.Save(world, Path.Combine(WorldManager.SavesFolderDir, serverConfig.SaveName));
+            bool savedSuccessfully = worldPersistence.Save(world, Path.Combine(OldWorldManager.SavesFolderDir, serverConfig.SaveName));
             if (savedSuccessfully && !string.IsNullOrWhiteSpace(serverConfig.PostSaveCommandPath))
             {
                 try
